requirements.txt
appdirs==1.4.1
Django==1.10.5
django-cors-headers==2.0.2
djangorestframework==3.5.4
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.1.10
six==1.10.0
